修正 bf16 → numpy 失敗（Blackwell 5090/B200 autocast 預設 bf16）
numpy 不支援 bfloat16；Blackwell GPU (5090 / B200) 的 autocast 在
某些路徑會輸出 bf16，到了 .cpu().numpy() 就炸：
    TypeError: Got unsupported ScalarType BFloat16

修正 4 處 .detach().cpu().numpy() → .detach().float().cpu().numpy()：
- TTS/tts/models/vits.py            (_log 中 sample_voice)
- TTS/vocoder/utils/generic_utils.py (plot_results y_hat / y)
- TTS/vocoder/models/gan.py          (train_log sample_voice)
- TTS/vocoder/models/wavegrad.py     (test_log sample_voice)

.float() 在 numpy 之前把 bf16/fp16 都升回 fp32，對非 bf16 tensor
是 no-op，安全。

+++ TTS/tts/models/vits.py
@@ -1338,7 +1338,9 @@ class Vits(BaseTTS):
         y_hat = outputs[1]["model_outputs"]
         y = outputs[1]["waveform_seg"]
         figures = plot_results(y_hat, y, ap, name_prefix)
-        sample_voice = y_hat[0].squeeze(0).detach().cpu().numpy()
+        # `.float()` upcasts bf16/fp16 → fp32 so .numpy() works on autocast
+        # outputs (numpy lacks bf16 entirely; Blackwell autocast emits bf16).
+        sample_voice = y_hat[0].squeeze(0).detach().float().cpu().numpy()
         audios = {f"{name_prefix}/audio": sample_voice}
 
         alignments = outputs[1]["alignments"]
