+++ TTS/vocoder/models/gan.py
@@ -200,7 +200,8 @@ class GAN(BaseVocoder):
         y_hat = outputs[0]["model_outputs"] if self.train_disc else outputs[1]["model_outputs"]
         y = batch["waveform"]
         figures = plot_results(y_hat, y, ap, name)
-        sample_voice = y_hat[0].squeeze(0).detach().cpu().numpy()
+        # `.float()` upcasts bf16/fp16 → fp32 so .numpy() works on autocast outputs
+        sample_voice = y_hat[0].squeeze(0).detach().float().cpu().numpy()
         audios = {f"{name}/audio": sample_voice}
         return figures, audios
